+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Here are a few common issues.
 ### Changed
 
 - Update to [TypeScript 3.7](https://devblogs.microsoft.com/typescript/announcing-typescript-3-7/).
+- There was an error when calling commands `testString` and `showVersion`.
 
 ## [0.1.0] - 2019-11-05